test(davinci): cover S2 source-map compatibility
